for #254: start switching to jest
