+++ package.json
@@ -34,10 +34,13 @@
     "eslint-plugin-node": "^6.0.1",
     "faucet": "^0.0.1",
     "htmllint-cli": "^0.0.6",
+    "jest": "^23.5.0",
     "node-mocks-http": "^1.7.0",
     "npm-run-all": "^4.1.3",
     "nsp": "^3.2.1",
     "nyc": "^11.8.0",
+    "proxyquire": "^2.0.1",
+    "sinon": "^6.1.5",
     "stylelint": "^9.2.0",
     "stylelint-config-standard": "^18.2.0",
     "tape": "^4.9.1",
